requirements.txt
# local package
-e .

# external requirements
click
Sphinx
coverage
awscli
flake8
python-dotenv>=0.5.1
tensorflow==2.1.0
lightgbm
setuptools>=41.0.0
docutils>=0.10
scipy
xgboost==0.90
cvxopt
jupyter_contrib_nbextensions
sklearn-pandas
keras
h2o
optuna
tpot